Write a function that multiplies two integers.

+++ 0x04-more_functions_nested_loops/2-mul.c
@@ -12,7 +12,7 @@
  * Return: 0
  */
 
-int mul(int a, int b)
+long int mul(int a, int b)
 {
 return (a * b);
 }
